requirements.txt
asgiref==3.5.2
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
djangorestframework==3.13.1
Flask==1.1.1
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.0
PySocks==1.7.1
pytz==2022.2.1
requests==2.22.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tweepy==3.8.0
twitivity==0.3.5
tzdata==2022.2
urllib3==1.25.11
Werkzeug==2.2.2
